+++ src/error.rs
@@ -5,6 +5,12 @@ type FallbackError = Box<dyn std::error::Error + Send + Sync>;
 
 #[derive(Debug, Error)]
 pub enum Error {
+    /// `NotFound` has been deprecated, please don't check this error in your
+    /// code (use "_" to ommit it if you are using match statement).
+    ///
+    /// The crate won't return this error anymore. It will be deleted in next
+    /// major version.
+    #[deprecated(since = "1.5.0", note = "won't return this error anymore")]
     #[error("exif/metadata not found")]
     NotFound,
 
@@ -13,6 +19,9 @@ pub enum Error {
 
     #[error("invalid entry; {0}")]
     InvalidEntry(FallbackError),
+
+    #[error("parsed entry result has been taken")]
+    EntryHasBeenTaken,
 }
 
 use Error::*;